new file - dungeon_move.c and clean-up
