+++ ld_script.txt
@@ -235,8 +235,7 @@ SECTIONS {
         src/code_804AFAC.o(.text);
         asm/code_804FD30.o(.text);
         src/dungeon_message.o(.text);
-        src/charge_move.o(.text);
-        src/move_util.o(.text);
+        src/dungeon_move.o(.text);
         src/move_actions.o(.text);
         src/status_actions.o(.text);
         src/move_checks.o(.text);
@@ -557,7 +556,7 @@ SECTIONS {
         src/dungeon_pokemon_attributes.o(.rodata);
         src/called_move_data.o(.rodata);
         data/data_80F59C8.o(.rodata);
-        src/charge_move.o(.rodata);
+        src/dungeon_move.o(.rodata);
         data/data_8106A4C.o(.rodata);
         src/code_806CD90.o(.rodata);
         src/type_effectiveness.o(.data);
