+++ sym_ewram.txt
@@ -392,47 +392,8 @@ gUnknown_202F1E0: /* 202F1E0 (sub_80511F0) */
 gUnknown_202F1E1: /* 202F1E1 (sub_80511F0 - sub_8051438) */
 	.space 0x7
 
-gUnknown_202F1E8: /* 202F1E8 (sub_80521D0 - sub_80523A8) */
-	.space 0x8
-
-gUnknown_202F1F0: /* 202F1F0 (sub_8052FB8) */
-	.space 0x8
-
-gUnknown_202F1F8: /* 202F1F8 (sub_80532B4 - sub_8053540) */
-	.space 0x4
-gUnknown_202F1FC: /* 202F1FC (sub_80532B4 - sub_8053540) */
-	.space 0x4
-
-gUnknown_202F200: /* 202F200 (sub_805363C) */
-	.space 0x8
-
-gUnknown_202F208: /* 202F208 (sub_8053704 - sub_8067904) */
-	.space 0x4
-gUnknown_202F20C: /* 202F20C (sub_8053704 - sub_8055FA0) */
-	.space 0x4
-gUnknown_202F210: /* 202F210 (sub_8055FA0 - sub_805A210) */
-	.space 0x4
-gUnknown_202F214: /* 202F214 (sub_8055FA0 - sub_8059988) */
-	.space 0x4
-gUnknown_202F218: /* 202F218 (sub_8055FA0 - sub_8059988) */
-	.space 0x1
-gUnknown_202F219: /* 202F219 (sub_8055FA0 - sub_80588B8) */
-	.space 0x1
-gUnknown_202F21A: /* 202F21A (sub_8055FA0 - sub_8058A18) */
-	.space 0x2
-gUnknown_202F21C: /* 202F21C (sub_8055FA0 - sub_805768C) */
-	.space 0x4
-gUnknown_202F220: /* 202F220 (sub_8055FA0 - sub_805768C) */
-	.space 0x1
-gUnknown_202F221: /* 202F221 (sub_8055A00 - sub_8072CF4) */
-	.space 0x1
-gUnknown_202F222: /* 202F222 (sub_8053704 - sub_8072CF4) */
-	.space 0x2
-gUnknown_202F224: /* 202F224 (sub_8040DA0 - sub_805B264) */
-	.space 0x4
-gMetronomeCalledArrayId: /* 202F228 (sub_8055FA0 - sub_805B618) */
-	.space 0x4
-
+    .include "src/dungeon_message.o"
+    .include "src/dungeon_move.o"
     .include "src/code_805D8C8_1.o"
 
 gUnknown_202F268: /* 202F268 (sub_8062500) */
